move OpenReacModifications to NetworkModifications
