+++ open-reac/src/main/java/com/powsybl/openreac/parameters/output/network/NetworkModifications.java
@@ -8,6 +8,10 @@ package com.powsybl.openreac.parameters.output.network;
 
 import com.powsybl.ampl.executor.AmplOutputFile;
 import com.powsybl.iidm.modification.GeneratorModification;
+import com.powsybl.iidm.modification.ShuntCompensatorModification;
+import com.powsybl.iidm.modification.StaticVarCompensatorModification;
+import com.powsybl.iidm.modification.VscConverterStationModification;
+import com.powsybl.iidm.modification.tapchanger.RatioTapPositionModification;
 import com.powsybl.iidm.network.Network;
 
 import java.util.List;
@@ -19,16 +23,40 @@ import java.util.List;
 public class NetworkModifications {
 
     private final GeneratorNetworkOutput generatorNetworkOutput;
+    private final ShuntCompensatorNetworkOutput shuntsOutput;
+    private final VscNetworkOutput vscOutput;
+    private final SvcNetworkOutput svcOutput;
+    private final TapNetworkOutput tapOutput;
 
     public NetworkModifications(Network network) {
         generatorNetworkOutput = new GeneratorNetworkOutput(network);
+        shuntsOutput = new ShuntCompensatorNetworkOutput(network);
+        vscOutput = new VscNetworkOutput(network);
+        svcOutput = new SvcNetworkOutput(network);
+        tapOutput = new TapNetworkOutput();
     }
 
     public List<AmplOutputFile> getOutputFiles() {
-        return List.of(generatorNetworkOutput);
+        return List.of(generatorNetworkOutput, shuntsOutput, vscOutput, svcOutput, tapOutput);
     }
 
     public List<GeneratorModification> getGeneratorModifications() {
         return generatorNetworkOutput.getModifications();
+    }
+
+    public List<ShuntCompensatorModification> getShuntModifications() {
+        return shuntsOutput.getModifications();
+    }
+
+    public List<VscConverterStationModification> getVscModifications() {
+        return vscOutput.getModifications();
+    }
+
+    public List<StaticVarCompensatorModification> getSvcModifications() {
+        return svcOutput.getModifications();
+    }
+
+    public List<RatioTapPositionModification> getTapModifications() {
+        return tapOutput.getModifications();
     }
 }
